Seperate Files
Seperated the classes into differnt files.
