+++ src/AuthMePE/AuthMePE.php
@@ -878,406 +878,3 @@ class AuthMePE extends PluginBase implements Listener{
 	}
 	
 }
-
-/* OTHER
- * ██████  █      ███   █████ █████
- * █       █     █   █  █     █     
- * █       █     █████  █████ █████
- * █       █     █   █      █     █ 
- * ██████  █████ █   █  █████ █████
- */
-
-namespace AuthMePE;
-
-use pocketmine\event\plugin\PluginEvent;
-use AuthMePE\AuthMePE;
-
-abstract class BaseEvent extends PluginEvent{
-	
-	public function __construct(AuthMePE $plugin){
-		$this->plugin = $plugin;
-		parent::__construct($plugin);
-	}
-}
-
-namespace AuthMePE;
-
-use pocketmine\Player;
-use pocketmine\event\Cancellable;
-
-use AuthMePE\AuthMePE;
-use AuthMePE\BaseEvent;
-
-class PlayerAuthEvent extends BaseEvent implements Cancellable{
-	public static $handlerList = null;
-	
-	private $player;
-	private $method;
-	
-	const PASSWORD = 0;
-	const IP = 1;
-	const PERMISSION = 2;
-	const SESSION = 3;
-	const COMMAND = 4;
-	
-	public function __construct(AuthMePE $plugin, Player $player, $method){
-		$this->player = $player;
-		$this->method = $method;
-		parent::__construct($plugin);
-	}
-	
-	public function getPlayer(){
-		return $this->player;
-	}
-	
-	public function getMethod(){
-		return $this->method;
-	}
-	
-	public function getIp(){
-		return $this->player->getAddress();
-	}
-}
-
-namespace AuthMePE;
-
-use pocketmine\event\Cancellable;
-use pocketmine\Player;
-
-use AuthMePE\AuthMePE;
-use AuthMePE\BaseEvent;
-
-class PlayerLogoutEvent extends BaseEvent implements Cancellable{
-	public static $handlerList = null;
-	
-	private $player;
-	
-	public function __construct(AuthMePE $plugin, Player $player){
-		$this->player = $player;
-		parent::__construct($plugin);
-	}
-	
-	public function getPlayer(){
-		return $this->player;
-	}
-}
-
-namespace AuthMePE;
-
-use pocketmine\Player;
-use pocketmine\event\Cancellable;
-
-use AuthMePE\AuthMePE;
-use AuthMePE\BaseEvent;
-
-class PlayerRegisterEvent extends BaseEvent implements Cancellable{
-	public static $handlerList = null;
-	
-	private $player;
-	
-	public function __construct(AuthMePE $plugin, Player $player){
-		$this->player = $player;
-		parent::__construct($plugin);
-	}
-	
-	public function getPlayer(){
-		return $this->player;
-	}
-}
-
-namespace AuthMePE;
-
-use pocketmine\Player;
-use pocketmine\event\Cancellable;
-
-use AuthMePE\AuthMePE;
-use AuthMePE\BaseEvent;
-
-class PlayerUnregisterEvent extends BaseEvent implements Cancellable{
-	public static $handlerList = null;
-	
-	private $player;
-	
-	public function __construct(AuthMePE $plugin, Player $player){
-		$this->player = $player;
-		parent::__construct($plugin);
-	}
-	
-	public function getPlayer(){
-		return $this->player;
-	}
-}
-
-namespace AuthMePE;
-
-use pocketmine\Player;
-use pocketmine\event\Cancellable;
-
-use AuthMePE\AuthMePE;
-use AuthMePE\BaseEvent;
-
-class PlayerChangePasswordEvent extends BaseEvent implements Cancellable{
-	public static $handlerList = null;
-	
-	private $player;
-	
-	public function __construct(AuthMePE $plugin, Player $player){
-		$this->player = $player;
-		parent::__construct($plugin);
-	}
-	
-	public function getPlayer(){
-		return $this->player;
-	}
-}
-
-namespace AuthMePE;
-
-use pocketmine\Player;
-use pocketmine\event\Cancellable;
-
-use AuthMePE\AuthMePE;
-use AuthMePE\BaseEvent;
-
-class PlayerAddEmailEvent extends BaseEvent implements Cancellable{
-	public static $handlerList = null;
-	
-	private $player;
-	
-	public function __construct(AuthMePE $plugin, Player $player, $email){
-		$this->player = $player;
-		$this->email = $email;
-		parent::__construct($plugin);
-	}
-	
-	public function getPlayer(){
-		return $this->player;
-	}
-	
-	public function getEmail(){
-		return $this->email;
-	}
-}
-
-namespace AuthMePE;
-
-use pocketmine\Player;
-
-use AuthMePE\AuthMePE;
-use AuthMePE\BaseEvent;
-
-class PlayerLoginTimeoutEvent extends BaseEvent{
-	public static $handlerList = null;
-	
-	private $player;
-	
-	public function __construct(AuthMePE $plugin, Player $player){
-		$this->player = $player;
-		parent::__construct($plugin);
-	}
-	
-	public function getPlayer(){
-		return $this->player;
-	}
-}
-
-namespace AuthMePE;
-
-use pocketmine\Player;
-use pocketmine\event\Cancellable;
-
-use AuthMePE\AuthMePE;
-use AuthMePE\BaseEvent;
-
-class PlayerAuthSessionStartEvent extends BaseEvent implements Cancellable{
-	public static $handlerList = null;
-	
-	private $player;
-	
-	public function __construct(AuthMePE $plugin, Player $player){
-		$this->player = $player;
-		parent::__construct($plugin);
-	}
-	
-	public function getPlayer(){
-		return $this->player;
-	}
-}
-
-namespace AuthMePE;
-
-use pocketmine\Player;
-
-use AuthMePE\AuthMePE;
-use AuthMePE\BaseEvent;
-
-class PlayerAuthSessionExpireEvent extends BaseEvent{
-	public static $handlerList = null;
-	
-	private $player;
-	
-	public function __construct(AuthMePE $plugin, Player $player){
-		$this->player = $player;
-		parent::__construct($plugin);
-	}
-	
-	public function getPlayer(){
-		return $this->player;
-	}
-}
-
-namespace AuthMePE;
-
-use pocketmine\scheduler\PluginTask;
-use AuthMePE\AuthMePE;
-use pocketmine\utils\TextFormat;
-use pocketmine\level\sound\PopSound;
-
-class Task extends PluginTask{
-	public $plugin;
-	
-	public function __construct(AuthMePE $plugin){
-		$this->plugin = $plugin;
-		parent::__construct($plugin);
-	}
-	
-	public function onRun($tick){
-		foreach($this->plugin->getServer()->getOnlinePlayers() as $p){
-			if($this->plugin->isRegistered($p) && !$this->plugin->isLoggedIn($p)){
-				$p->sendMessage("Please type your password in chat to login.");
-				$p->sendPopup(TextFormat::GOLD."Welcome ".TextFormat::AQUA.$p->getName().TextFormat::GREEN."\nPlease login to play!");
-				$p->getLevel()->addSound(new PopSound($p), $this->plugin->getServer()->getOnlinePlayers());
-			}
-		}
-	}
-}
-
-namespace AuthMePE;
-
-use pocketmine\scheduler\PluginTask;
-use pocketmine\utils\TextFormat;
-
-use AuthMePE\AuthMePE;
-use AuthMePE\PlayerLoginTimeoutEvent;
-
-class Task2 extends PluginTask{
-	public $plugin;
-	
-	public function __construct(AuthMePE $plugin, $player){
-		$this->plugin = $plugin;
-		$this->player = $player;
-		parent::__construct($plugin);
-	}
-	
-	public function onRun($tick){
-			if(!$this->plugin->isLoggedIn($this->player) || !$this->plugin->isRegistered($this->player)){
-				$this->plugin->getServer()->getPluginManager()->callEvent(new PlayerLoginTimeoutEvent($this->plugin, $this->player));
-				$this->player->kick(TextFormat::RED."Time out!");
-				$this->plugin->getServer()->getScheduler()->cancelTask($this->getTaskId());
-			}
-	}
-}
-
-namespace AuthMePE;
-
-use pocketmine\scheduler\PluginTask;
-use pocketmine\Player;
-
-use AuthMePE\AuthMePE;
-
-class SessionTask extends PluginTask{
-	public $plugin;
-	public $player;
-	
-	public function __construct(AuthMePE $plugin, $player){
-		$this->plugin = $plugin;
-		$this->player = $player;
-		parent::__construct($plugin);
-	}
-	
-	public function onRun($tick){
-		$this->plugin->closeSession($this->player);
-		$this->plugin->getServer()->getScheduler()->cancelTask($this->getTaskId());
-	}
-}
-
-namespace AuthMePE;
-
-use pocketmine\scheduler\PluginTask;
-use pocketmine\Player;
-use pocketmine\level\sound\BatSound;
-use pocketmine\level\sound\LaunchSound;
-
-use AuthMePE\AuthMePE;
-
-class SoundTask extends PluginTask{
-	public $plugin;
-	public $player;
-	public $type;
-	
-	public function __construct(AuthMePE $plugin, $player, $type){
-		$this->plugin = $plugin;
-		$this->player = $player;
-		$this->type = $type;
-		parent::__construct($plugin);
-	}
-	
-	public function onRun($tick){
-		switch($this->type){
-			case 1:
-			  $this->player->getLevel()->addSound(new BatSound($this->player), $this->player->getServer()->getOnlinePlayers());
-			break;
-			case 2:
-			  $this->player->getLevel()->addSound(new LaunchSound($this->player), $this->player->getServer()->getOnlinePlayers());
-			break;
-		}
-	}
-}
-
-namespace AuthMePE;
-
-use pocketmine\scheduler\PluginTask;
-use pocketmine\Player;
-
-use AuthMePE\AuthMePE;
-
-class UnbanTask extends PluginTask{
-	public $plugin;
-	public $player;
-	
-	public function __construct(AuthMePE $plugin, $player){
-		$this->plugin = $plugin;
-		$this->player = $player;
-		parent::__construct($plugin);
-	}
-	
-	public function onRun($tick){
-		$this->plugin->unban($this->player->getName());
-		$this->plugin->getServer()->getScheduler()->cancelTask($this->getTaskId());
-	}
-}
-
-namespace AuthMePE;
-
-use pocketmine\scheduler\PluginTask;
-
-use AuthMePE\AuthMePE;
-
-class TokenDeleteTask extends PluginTask{
-	public $plugin;
-	
-	public function __construct(AuthMePE $plugin){
-		$this->plugin = $plugin;
-		parent::__construct($plugin);
-	}
-	
-	public function onRun($tick){
-		if($this->plugin->getToken() !== null){
-		  $this->plugin->delToken();
-		  $this->plugin->getLogger()->info("Last generated token has expired!  Please generate a new one!");
-		}
-		$this->plugin->getServer()->getScheduler()->cancelTask($this->getTaskId());
-	}
-}
-
-?>
